pyproject.toml
[tool.poetry]
name = "ska-sdp-dataproduct-metadata"
version = "0.0.0"
description = "SKA SDP Data Product Metadata"
authors = ["SKA SDP Developers"]
repository = "https://gitlab.com/ska-telescope/sdp/ska-sdp-dataproduct-metadata"
documentation = "https://developer.skao.int/projects/ska-sdp-scripting/en/latest/"
license = "BSD-3-Clause"

[[tool.poetry.source]]
name = "ska"
url = "https://artefact.skao.int/repository/pypi-internal/simple"

[tool.poetry.dependencies]
python = "^3.9"
ska-ser-logging = "^0.4.1"
pyyaml = "^6.0"
ska-sdp-config = "^0.4.2"

[tool.poetry.dev-dependencies]
isort = "^5.10.1"
black = "^22.8.0"
flake8 = "^5.0.4"
pylint = "^2.15.2"
pylint-junit = "^0.3.2"
pytest = "^7.1.3"
pytest-cov = "^3.0.0"
build = "^0.8.0"

[tool.pylint.similarities]
min-similarity-lines = 10

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
